streamlit_app/app/app_requirements.txt
python-dotenv
pandas==2.2.3
plotly==5.24.1
requests==2.32.3
streamlit==1.40.2
streamlit_searchbox==0.1.19